Move invisible_text test into test_read.py
